nlp-student-kit/requirements.txt
torch>=2.1,<3.0
transformers>=4.40,<5.0
peft>=0.10,<1.0
tokenizers>=0.15
pyyaml>=6.0
tqdm>=4.66
openai>=1.20
